+++ example/ubiq_sample.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+const path = require('path');
 // Subject to the foregoing terms and conditions, Ubiq hereby grants to You, at
 // no cost, a perpetual, worldwide, non-exclusive, royalty-free, irrevocable
 // (except as stated herein) license to the Software, including all right to
@@ -29,211 +31,201 @@
 // REFLECT AN EQUITABLE ALLOCATION OF RISK BETWEEN THE PARTIES AND THAT IN
 // THEIR ABSENCE THE TERMS OF THIS LICENSE WOULD BE SUBSTANTIALLY DIFFERENT.
 
-const ubiq = require('ubiq-security')
+const { argv } = require('yargs');
+const ubiq = require('ubiq-security');
+const pkginfo = require('./package.json');
 
-const {argv} = require('yargs')
-var fs = require('fs');
-const util = require('util');
+function main() {
+  const credentials_file = argv.c;
+  const input_file = argv.i;
+  const outfile = argv.o;
+  let input_data = '';
+  const profile = argv.P;
 
-const credentials_file = argv.c
-const input_file = argv.i
-var outfile = argv.o
-var input_data = ''
-var profile = argv.P
+  // Blocks of 1 MiB
+  const BLOCK_SIZE = 1024 * 1024;
 
-// Blocks of 1 MiB
-const BLOCK_SIZE = 1024 * 1024
+  // Allow simple encryption / decryption for files less than 50 MiB
+  const MAX_SIMPLE_SIZE = 1025 * 1024 * 50;
 
-// Allow simple encryption / decryption for files less than 50 MiB
-MAX_SIMPLE_SIZE = 1025 * 1024 * 50
+  let process_mode = '';
+  let mode = '';
 
-var process_mode = ''
-var mode = ''
+  if (argv.V) {
+    console.log(`${argv.$0}: version: ${pkginfo.version}`);
+    return;
+  }
 
-if (argv.V) {
-  var pkginfo = require('./package.json')
-  console.log(argv.$0 + ": version: " + pkginfo.version)
-  return
-}
+  let invalid_option = false;
 
+  if (argv.e) {
+    process_mode = 'encrypt';
+  } else if (argv.d) {
+    process_mode = 'decrypt';
+  }
 
+  if (argv.s) {
+    mode = 'simple';
+  } else if (argv.p) {
+    mode = 'piecewise';
+  }
 
-invalid_option = false
+  if (!process_mode) {
+    console.log('Please provide a valid option');
+    invalid_option = true;
+  }
 
-if(argv.e){
-  process_mode = 'encrypt'
-}else if(argv.d){
-  process_mode = 'decrypt'
-}
+  if (!mode) {
+    console.log('Please provide a valid mode');
+    invalid_option = true;
+  }
 
-if(argv.s){
-  mode = 'simple'
-}else if(argv.p){
-  mode = 'piecewise'
-}
+  function display_prompt() {
+    console.log();
+    console.log('usage: ubiq_sample.js [-h] [-V] [-e] [-d] [-s] [-p] -i INFILE -o OUTFILE');
+    console.log('                      [-c CREDENTIALS] [-P PROFILE]');
+  }
 
+  function display_help() {
+    console.log();
+    console.log('  Sample application to provide examples of using the Ubiq Platform Node.js Client Library');
+    console.log();
+    console.log('  Created by Ubiq Security, Inc.');
+    console.log('  Copyright 2020- Ubiq Security, Inc., All rights reserved.');
+    console.log();
+    console.log('  Distributed on an "AS IS" basis without warranties');
+    console.log('  or conditions of any kind, either express or implied.');
+    console.log();
+    console.log('USAGE');
+    console.log();
+    console.log('optional arguments:');
+    console.log('-h               Show this help message and exit');
+    console.log('-V               Show program version number and exit');
+    console.log('-e               Encrypt the contents of the input file and write the');
+    console.log('                 results to output file');
+    console.log('-d               Decrypt the contents of the input file and write the');
+    console.log('                 results to output file');
+    console.log('-s,              Use the simple encryption / decryption interfaces');
+    console.log('-p,              Use the piecewise encryption / decryption interfaces');
+    console.log('-i INFILE,       Set input file name');
+    console.log('-o OUTFILE,      Set output file name');
+    console.log('-c CREDENTIALS   Set the file name with the API credentials');
+    console.log('                 (default: ~/.ubiq/credentials[.json]) ');
+    console.log('-P PROFILE       Identify the profile within the credentials file');
+    console.log('                 (default: default)');
+  }
 
-if(!process_mode){
-  console.log('Please provide a valid option')
-  invalid_option = true;
-}
+  if (argv.h) {
+    display_prompt();
+    display_help();
+    return;
+  }
 
-if(!mode){
-  console.log('Please provide a valid mode')
-  invalid_option = true;
-}
+  if (invalid_option) {
+    display_prompt();
+    return;
+  }
 
-function display_prompt() {
-  console.log()
-  console.log('usage: ubiq_sample.js [-h] [-V] [-e] [-d] [-s] [-p] -i INFILE -o OUTFILE')
-  console.log('                      [-c CREDENTIALS] [-P PROFILE]')
-}
+  const credentials = new ubiq.ConfigCredentials(credentials_file, profile);
 
-function display_help() {
-  console.log()
-  console.log('  Sample application to provide examples of using the Ubiq Platform Node.js Client Library')
-  console.log()
-  console.log('  Created by Ubiq Security, Inc.')
-  console.log('  Copyright 2020- Ubiq Security, Inc., All rights reserved.')
-  console.log()
-  console.log('  Distributed on an "AS IS" basis without warranties')
-  console.log('  or conditions of any kind, either express or implied.')
-  console.log()
-  console.log('USAGE')
-  console.log()
-  console.log('optional arguments:')
-  console.log('  -h             Show this help message and exit')
-  console.log('  -V             Show program version number and exit')
-  console.log('  -e             Encrypt the contents of the input file and write the')
-  console.log('                 results to output file')
-  console.log('  -d             Decrypt the contents of the input file and write the')
-  console.log('                 results to output file')
-  console.log('-s,              Use the simple encryption / decryption interfaces')
-  console.log('-p,              Use the piecewise encryption / decryption interfaces')
-  console.log('-i INFILE,       Set input file name')
-  console.log('-o OUTFILE,      Set output file name')
-  console.log('-c CREDENTIALS   Set the file name with the API credentials')
-  console.log('                 (default: ~/.ubiq/credentials[.json]) ')
-  console.log('  -P PROFILE     Identify the profile within the credentials file')
-  console.log('                 (default: default)')
-}
+  // Test to see if the credentials have been found and loaded properly
+  if (credentials.access_key_id === undefined
+    || credentials.secret_signing_key === undefined
+    || credentials.secret_crypto_access_key === undefined) {
+    console.log('  Unable to load credentials file properly.');
+    console.log('  Check credentials file pathname and selected profile');
+    return;
+  }
 
-if (argv.h) {
-  display_prompt()
-  display_help()
-  return
-}
+  function getStuff(infile) {
+    return fs.readFileSync(path.join(__dirname, infile), 'utf-8');
+  }
 
-if (invalid_option) {
-  display_prompt()
-  return
-}
+  async function getData(infile) {
+    const input = await getStuff(infile);
+    return input;
+  }
 
-let credentials = new ubiq.ConfigCredentials(credentials_file, profile)
+  async function simpleEncrypt() {
+    input_data = await getData(input_file);
+    const encrypted_res = await ubiq.encrypt(credentials, input_data);
 
-// Test to see if the credentials have been found and loaded properly
-if (credentials.access_key_id == undefined ||
-    credentials.secret_signing_key == undefined ||
-    credentials.secret_crypto_access_key == undefined) {
-  console.log ("  Unable to load credentials file properly.")
-  console.log ("  Check credentials file pathname and selected profile")
-  return
-}
-
-const readFile = util.promisify(fs.readFile);
-
-function getStuff(infile) {
-  return readFile(infile);
-}
-
-async function getData(infile){
-  input = await getStuff(infile)
-  return input
-}
-
-async function simpleEncrypt(){
-  input_data = await getData(input_file)
-  let encrypted_res = await ubiq.encrypt(credentials, input_data)
-
-  fs.writeFile(outfile, encrypted_res, function (err) {
-    if (err) throw err;
-  });
-}
-
-async function simpleDecrypt(){
-  input_data = await getData(input_file)
-
-  let decrypted = await ubiq.decrypt(credentials, input_data)
-
-  let bf = Buffer.from(decrypted, 'binary')
-
-  fs.writeFile(outfile, bf ,function (err) {
-    if (err) throw err;
-  });
-}
-
-async function pieceWiseEncrypt(uses){
-
-  let enc = await new ubiq.Encryption(credentials, uses);
-  // This returns the packed byte string
-  let begin_res = enc.begin()
-  var ws = fs.createWriteStream(outfile, {encoding: 'binary'});
-  ws.write(begin_res)
-  var readStream = fs.createReadStream(input_file,{ highWaterMark: BLOCK_SIZE  });
-  readStream.on('data', function(chunk) {
-    res = enc.update(chunk)
-    ws.write(res)
-  }).on('end', function() {
-      // TO indicate all the parsing has been complete
-      ws.write(enc.end())
-      ws.close()
-      enc.close()
+    fs.writeFileSync(outfile, encrypted_res, (err) => {
+      if (err) throw err;
     });
-}
-
-async function pieceWiseDecrypt(){
-
-  let dec = new ubiq.Decryption(credentials)
-  let begin_res = dec.begin()
-  var ws = fs.createWriteStream(outfile, {encoding: 'binary'});
-  ws.write(begin_res)
-  var readStream = fs.createReadStream(input_file,{ highWaterMark: BLOCK_SIZE });
-  readStream.on('data', async function(chunk) {
-    readStream.pause()
-    await dec.update(chunk).then(function(response){
-      if(response){
-        ws.write(response)
-      }
-    })
-    readStream.resume()
-  }).on('end', async function() {
-      ws.write(dec.end())
-      ws.close()
-      dec.close()
-  });
-
-}
-
-let stats = fs.statSync(input_file)
-var fileSizeInBytes = stats["size"]
-if (mode == 'simple' && fileSizeInBytes > MAX_SIMPLE_SIZE) {
-	console.log ("NOTE: This is only for demonstration purposes and is designed to work on memory")
-    console.log ("      constrained devices.  Therefore, this sample application will switch to")
-    console.log ("      the piecewise APIs for files larger than " + MAX_SIMPLE_SIZE + " bytes in order to reduce")
-    console.log ("      excessive resource usages on resource constrained IoT devices")
-    mode = 'piecewise'
-}
-
-if(mode == 'simple'){
-  if(process_mode == 'encrypt'){
-    simpleEncrypt()
-  }else{
-    simpleDecrypt()
+    console.log(`done: ${outfile}`);
   }
-}else{
-  if(process_mode == 'encrypt'){
-    pieceWiseEncrypt(1)
-  }else{
-    pieceWiseDecrypt()
+
+  async function simpleDecrypt() {
+    input_data = await getData(input_file);
+
+    const decrypted = await ubiq.decrypt(credentials, input_data);
+    const bf = Buffer.from(decrypted, 'binary');
+
+    fs.writeFile(outfile, bf, (err) => {
+      if (err) throw err;
+    });
+  }
+
+  async function pieceWiseEncrypt(uses) {
+    const enc = await new ubiq.Encryption(credentials, uses);
+    // This returns the packed byte string
+    const begin_res = enc.begin();
+    const ws = fs.createWriteStream(outfile, { encoding: 'binary' });
+    ws.write(begin_res);
+    const readStream = fs.createReadStream(input_file, { highWaterMark: BLOCK_SIZE });
+    readStream.on('data', (chunk) => {
+      const res = enc.update(chunk);
+      ws.write(res);
+    }).on('end', () => {
+      // TO indicate all the parsing has been complete
+      ws.write(enc.end());
+      ws.close();
+      enc.close();
+    });
+  }
+
+  async function pieceWiseDecrypt() {
+    const dec = new ubiq.Decryption(credentials);
+    const begin_res = dec.begin();
+    const ws = fs.createWriteStream(outfile, { encoding: 'binary' });
+    ws.write(begin_res);
+    const readStream = fs.createReadStream(input_file, { highWaterMark: BLOCK_SIZE });
+    readStream.on('data', async (chunk) => {
+      readStream.pause();
+      await dec.update(chunk).then((response) => {
+        if (response) {
+          ws.write(response);
+        }
+      });
+      readStream.resume();
+    }).on('end', async () => {
+      ws.write(dec.end());
+      ws.close();
+      dec.close();
+    });
+  }
+
+  const stats = fs.statSync(input_file);
+  const fileSizeInBytes = stats.size;
+  if (mode === 'simple' && fileSizeInBytes > MAX_SIMPLE_SIZE) {
+    console.log('NOTE: This is only for demonstration purposes and is designed to work on memory');
+    console.log('      constrained devices.  Therefore, this sample application will switch to');
+    console.log(`      the piecewise APIs for files larger than ${MAX_SIMPLE_SIZE} bytes in order to reduce`);
+    console.log('      excessive resource usages on resource constrained IoT devices');
+    mode = 'piecewise';
+  }
+
+  if (mode === 'simple') {
+    if (process_mode === 'encrypt') {
+      simpleEncrypt();
+    } else {
+      simpleDecrypt();
+    }
+  } else if (process_mode === 'encrypt') {
+    pieceWiseEncrypt(1);
+  } else {
+    pieceWiseDecrypt();
   }
 }
+main();