+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+##  2022-06-21
+* Added support for Format Preserving Encryption (FPE)
+* Added example program for Format Preserving Encryption (FPE)
+
 ## 1.0.5 - 2020-02-24
 * Added requirement for Node 12 or higher
 * Updated copyright notice and comments